+++ packages/core/src/utils/providers.ts
@@ -1,4 +1,5 @@
 import { Type } from '@ts-stack/di';
+import { LogFilter } from '../services/log-mediator';
 import { Logger, LoggerConfig } from '../types/logger';
 import { ServiceProvider } from '../types/mix';
 
@@ -42,7 +43,7 @@ import { ServiceProvider } from '../types/mix';
     .use(Plugin2)
     .method1()
     .method2()
-    .useValue(LoggerConfig, new LoggerConfig('trace'))
+    .useLogConfig({ level: 'trace' })
     .useClass(SomeService, ExtendedService)
  * ```
  * 
@@ -71,6 +72,15 @@ export class Providers {
     this.providers.push({ provide: Logger, useValue: useLogger });
     if (useConfig) {
       this.providers.push({ provide: LoggerConfig, useValue: useConfig });
+    }
+
+    return this;
+  }
+
+  useLogConfig(useConfig: LoggerConfig, logFilter?: LogFilter) {
+    this.providers.push({ provide: LoggerConfig, useValue: useConfig });
+    if (logFilter) {
+      this.providers.push({ provide: LogFilter, useValue: logFilter });
     }
 
     return this;